sporo zmian, poprawek, wersja bez swaggera

+++ BUSINESS/src/main/java/facebookapi/business/service/CommentService.java
@@ -2,16 +2,21 @@ package facebookapi.business.service;
 
 import facebookapi.business.IdChecker;
 import facebookapi.business.dto.CommentDto;
-import facebookapi.business.dto.NewCommentDto;
 import facebookapi.business.mappers.CommentMapper;
+import facebookapi.business.payload.request.NewCommentRequest;
+import facebookapi.business.security.services.UserDetailsImpl;
 import facebookapi.domain.entity.Comment;
 import facebookapi.domain.entity.Post;
 import facebookapi.domain.entity.User;
 import facebookapi.domain.repository.CommentRepository;
+import facebookapi.domain.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -20,19 +25,20 @@ public class CommentService {
     private final CommentRepository commentRepository;
     private final CommentMapper commentMapper;
     private final IdChecker idChecker;
+    private final UserService userService;
+    private final UserRepository userRepository;
 
-    public CommentDto saveComment(NewCommentDto newComment) {
-        idChecker.isUserAvailable(newComment.getUserId());
-        idChecker.isPostAvailable(newComment.getPostId());
+    public CommentDto saveComment(NewCommentRequest newComment) {
+        idChecker.checkPostAvailable(newComment.getPostId());
 
-        Comment comment = commentMapper.dtoToComment(newComment);
+        Comment comment = commentMapper.toComment(newComment);
         var savedComment = commentRepository.save(comment);
 
         return commentMapper.toCommentDto(savedComment);
     }
 
     public List<CommentDto> getCommentsByPostId(int postId) {
-        Post postAvailable = idChecker.isPostAvailable(postId);
+        Post postAvailable = idChecker.checkPostAvailable(postId);
         List<Comment> comments = commentRepository.findAllByPost(postAvailable);
 
         return commentMapper.toCommentsDto(comments);
@@ -46,14 +52,21 @@ public class CommentService {
     }
 
     public String deleteComment(int commentId) {
-        idChecker.isCommentAvailable(commentId);
-        commentRepository.deleteById(commentId);
+        Comment comment = idChecker.checkCommentAvailable(commentId);
+        int userId = comment.getUser().getUserId();
 
-        return "Komentarz o numerze Id: " + commentId + " zostal pomyslnie usuniety.";
+        int user_Id = userService.retrieveCurrentlyAuthenticatedUserId();
+        if (user_Id == userId) {
+            commentRepository.deleteById(commentId);
+
+            return "Komentarz o numerze Id: " + commentId + " zostal pomyslnie usuniety.";
+
+        } else
+            return "Error: Mozesz usunac tylko swoj komentarz!";
     }
 
     public List<CommentDto> getCommentsByUserId(int userId) {
-        User userAvailable = idChecker.isUserAvailable(userId);
+        User userAvailable = idChecker.checkUserAvailable(userId);
 
         List<Comment> userComments = commentRepository.findAllByUser(userAvailable);
         List<CommentDto> commentsDto = commentMapper.toCommentsDto(userComments);
@@ -61,13 +74,13 @@ public class CommentService {
         return commentsDto;
     }
 
-    public String deleteCommentsByUserId(int userId) {
-        User userAvailable = idChecker.isUserAvailable(userId);
-        List<Comment> userComments = commentRepository.findAllByUser(userAvailable);
+    public String deleteUserComments() {
+        int userId = userService.retrieveCurrentlyAuthenticatedUserId();
+        Optional<User> user = userRepository.findById(userId);
+
+        List<Comment> userComments = commentRepository.findAllByUser(user.get());
         commentRepository.deleteAll(userComments);
 
-        return "Usunieto wszystkie komentarze dodane przez uzytkownika o numerze Id: " + userId;
+            return "Usunieto wszystkie Twoje komentarze!";
     }
-
-
 }
