+++ BUSINESS/src/main/java/facebookapi/business/service/PostService.java
@@ -1,16 +1,22 @@
 package facebookapi.business.service;
 
 import facebookapi.business.IdChecker;
-import facebookapi.business.dto.NewPostDto;
+import facebookapi.business.payload.request.NewPostRequest;
 import facebookapi.business.dto.PostDto;
 import facebookapi.business.mappers.PostMapper;
+import facebookapi.business.security.services.UserDetailsImpl;
 import facebookapi.domain.entity.Post;
 import facebookapi.domain.entity.User;
 import facebookapi.domain.repository.PostRepository;
+import facebookapi.domain.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
+import java.lang.reflect.Field;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -19,16 +25,16 @@ public class PostService {
     private final PostRepository postRepository;
     private final PostMapper postMapper;
     private final IdChecker idChecker;
+    private final UserRepository userRepository;
+    private final UserService userService;
 
-    public PostDto savePost(NewPostDto newPostDto) {
-        idChecker.isUserAvailable(newPostDto.getUserId());
 
+    public PostDto savePost(NewPostRequest newPostRequest) {
 
-        Post newPost = postMapper.dtoToPost(newPostDto);
+        Post newPost = postMapper.toPost(newPostRequest);
         var savedPost = postRepository.save(newPost);
 
         return postMapper.toPostDto(savedPost);
-
     }
 
 
@@ -42,28 +48,40 @@ public class PostService {
         return postsDto;
     }
 
-    public String deletePost(int postId) {
-        idChecker.isPostAvailable(postId);
-        postRepository.deleteById(postId);
 
-        return "Post zostal pomyslnie usuniety";
+    public String deletePost(int postId) {
+        Post post = idChecker.checkPostAvailable(postId);
+        int user_Id = post.getUser().getUserId();
+
+        int userId = userService.retrieveCurrentlyAuthenticatedUserId();
+        if (user_Id == userId){
+            postRepository.deleteById(postId);
+
+            return "Post zostal pomyslnie usuniety.";
+
+        } else
+            return "Error: Mozesz usunac tylko swoj post!";
     }
 
+
     public List<PostDto> getUserPosts(int userId) {
-        User userAvailable = idChecker.isUserAvailable(userId);
+        User userAvailable = idChecker.checkUserAvailable(userId);
         List<Post> userPosts = postRepository.findByUser(userAvailable);
 
         return postMapper.toPostsDto(userPosts);
     }
 
-    public String deleteUserPosts(int userId) {
-        User userAvailable = idChecker.isUserAvailable(userId);
-        List<Post> userPosts = postRepository.findByUser(userAvailable);
+    public String deleteUserPosts() {
+        int user_Id = userService.retrieveCurrentlyAuthenticatedUserId();
+        Optional<User> user = userRepository.findById(user_Id);
 
-        userPosts.forEach(post -> {
-            postRepository.delete(post);
-        });
+        List<Post> userPosts = postRepository.findByUser(user.get());
 
-        return "Usunieto pomyslnie wszystkie posty uzytkownika o numerze Id: " + userId;
+            userPosts.forEach(post -> {
+                postRepository.delete(post);
+            });
+
+            return "Usunieto pomyslnie wszystkie Twoje posty!";
     }
+
 }
